[chore] Refactor TestContext and Apply Feedback from PR #1738 (#1834)
* Refactor: Apply feedback from previous PR #1738 to improve code readability and structure

* Moved the `TestContext` struct and its associated methods to a dedicated `test_context_test.go` file.

* Move `RunTestCases` out of `TestContext`

* Move `ensureResourceApplied` out of `TestContext`

* Move fetcher functions out of `TestContext`

* Move `NewResourceOptions` from the `resource_options_test.go` to the `test_context_test.go` file

* Move `createInstallPlan` and `createSubscription` to `TestContext`

Migrate the `createInstallPlan` and `createSubscription` functions to `TestContext` for better organization and clarity.

* Move `TestContext` struct and initialization method to `test_context_test.go`
