+++ tests/e2e/codeflare_test.go
@@ -32,5 +32,5 @@ func codeflareTestSuite(t *testing.T) {
 	}
 
 	// Run the test suite.
-	componentCtx.RunTestCases(t, testCases)
+	RunTestCases(t, testCases)
 }
