+++ tests/e2e/feastoperator_test.go
@@ -32,5 +32,5 @@ func feastOperatorTestSuite(t *testing.T) {
 	}
 
 	// Run the test suite.
-	componentCtx.RunTestCases(t, testCases)
+	RunTestCases(t, testCases)
 }
