+++ tests/e2e/modelmeshserving_test.go
@@ -33,5 +33,5 @@ func modelMeshServingTestSuite(t *testing.T) {
 	}
 
 	// Run the test suite.
-	componentCtx.RunTestCases(t, testCases)
+	RunTestCases(t, testCases)
 }
