+++ tests/e2e/ray_test.go
@@ -32,5 +32,5 @@ func rayTestSuite(t *testing.T) {
 	}
 
 	// Run the test suite.
-	componentCtx.RunTestCases(t, testCases)
+	RunTestCases(t, testCases)
 }
