+++ tests/e2e/trainingoperator_test.go
@@ -32,5 +32,5 @@ func trainingOperatorTestSuite(t *testing.T) {
 	}
 
 	// Run the test suite.
-	componentCtx.RunTestCases(t, testCases)
+	RunTestCases(t, testCases)
 }
