+++ tests/e2e/workbenches_test.go
@@ -34,5 +34,5 @@ func workbenchesTestSuite(t *testing.T) {
 	}
 
 	// Run the test suite.
-	componentCtx.RunTestCases(t, testCases)
+	RunTestCases(t, testCases)
 }
